Use `compose-with-values` macro instead of `compose`
This was suggested by @samth, @notjack and @benknoble, as a way both
to make composition using `~>` more performant, and as a step towards
making it interoperable with Typed Racket.

Currently, although standalone benchmarking shows a > 2x improvement,
our "smoke" nonlocal benchmark for composition seems to show a slight
drop in performance. Reason for the discrepancy unclear.